[BE] fix: rebase conflict 처리완료
- rebase 이후 에러 해결
- 누락코드 추가 (확인필요)
- createGroupUsingStack UsingTechStackService로 이동
