+++ server/src/domains/group/api/GroupController.ts
@@ -1,11 +1,10 @@
 import { UsingTechStackService } from '@domains/techstack/service/UsingTechStackService';
 import { ProfileService } from '@domains/user/service/ProfileService';
-import { TechStackService } from '@domains/techstack/service/TechStackService';
-import { Body, Controller, OnUndefined, Post } from 'routing-controllers';
+import { Body, Controller, Get, OnUndefined, Post } from 'routing-controllers';
 import { Inject, Service } from 'typedi';
-import { Group } from '../dto/groupDto';
 import { CreateGroupDto } from '../dto/CreateGroupDto';
 import { GroupService } from '../service/GroupService';
+import { Group } from '../models/Group';
 
 @Service()
 @Controller('/group')
@@ -21,8 +20,8 @@ export class GroupController {
 
   @Get('/')
   async getAll() {
-    const addedOtherGroupInfo = await this.addOtherGroupInfo(groups);
     const groups: Group[] = await this.groupService.getGroups();
+    const addedOtherGroupInfo = await this.addOtherGroupInfo(groups);
     return addedOtherGroupInfo;
   }
 
@@ -30,7 +29,7 @@ export class GroupController {
   @OnUndefined(200)
   async create(@Body() groupData: CreateGroupDto) {
     const createdGroup = await this.groupService.createGroup(groupData);
-    this.techStackService.createGroupUsingStack(createdGroup, groupData.usingTechStacks);
+    this.usingTechStackService.createGroupUsingStack(createdGroup, groupData.usingTechStacks);
   }
 
   addOtherGroupInfo(groups: Group[]) {
